Starting work on 1.6
