+++ part1/src/App.js
@@ -1,72 +1,14 @@
+import { useState } from 'react'
+
 const App = () => {
-
-  const Header = (props) => {
-    return (
-      <>
-        <h1>
-          {props.course}
-        </h1>
-      </>
-    )
-  }
-
-  const Part = (props) => {
-    return (
-      <>
-        <p>
-          {props.part} {props.exercises}
-        </p>
-      </>
-    )
-  }
-
-  const Content = (props) => {
-    return (
-      <>
-        <Part part={props.parts[0].name} exercises={props.parts[0].exercises} />
-        <Part part={props.parts[1].name} exercises={props.parts[1].exercises} />
-        <Part part={props.parts[2].name} exercises={props.parts[2].exercises} />
-      </>
-    )
-  }
-
-  const Total = (props) => {
-    const total = props.parts.map(parts => parts.exercises)
-                             .reduce(
-                              (accumulator, currentValue) => accumulator + currentValue, 0);
-    return (
-      <>
-        <p>
-          Number of exercises {total} 
-        </p>
-      </>
-    )
-  }
-
-  const course = {
-    name: 'Half Stack application development',
-    parts: [
-      {
-        name: 'Fundamentals of React',
-        exercises: 10
-      },
-      {
-        name: 'Using props to pass data',
-        exercises: 7
-      },
-      {
-        name: 'State of a component',
-        exercises: 14
-      }
-    ]
-  }
-
+  // save clicks of each button to its own state
+  const [good, setGood] = useState(0)
+  const [neutral, setNeutral] = useState(0)
+  const [bad, setBad] = useState(0)
 
   return (
     <div>
-      <Header course={course.name} />
-      <Content parts={course.parts} />
-      <Total parts={course.parts} />
+      code here
     </div>
   )
 }
